CI Environment (#1)
Set up a build environment inside a Docker container:

* Add `ci` subdirectory where the new stuff lives
* Add Makefile targets to build the container, build the code, and
output artifacts
* All dependencies _and_ the source code are packaged up in the
container, which means you can run the final `cargo build` without an
Internet connection (which the Makefile does).
* GitHub now runs this process during pull requests, pushes to `main`,
and pushes to release tags. It also attaches the produced artifacts to
the Actions run, so it can be downloaded by anyone who's interested.

With one caveat: I only have a Linux machine to work with, so I only
focused on the Linux build. I haven't touched macOS or Windows, and I
don't plan to.